Match int64 usage of ffmpeg.
Should't use int, since its size depends on architecture.

+++ format.go
@@ -416,7 +416,7 @@ func (this *FmtCtx) SetFlag(flag int) *FmtCtx {
 	return this
 }
 
-func (this *FmtCtx) SeekFile(ist *Stream, minTs, maxTs int, flag int) error {
+func (this *FmtCtx) SeekFile(ist *Stream, minTs, maxTs int64, flag int) error {
 	if ret := int(C.avformat_seek_file(this.avCtx, C.int(ist.Index()), C.int64_t(0), C.int64_t(minTs), C.int64_t(maxTs), C.int(flag))); ret < 0 {
 		return errors.New(fmt.Sprintf("Error creating output context: %s", AvError(ret)))
 	}
@@ -424,13 +424,13 @@ func (this *FmtCtx) SeekFile(ist *Stream, minTs, maxTs int, flag int) error {
 	return nil
 }
 
-func (this *FmtCtx) SeekFrameAt(sec int, streamIndex int) error {
+func (this *FmtCtx) SeekFrameAt(sec int64, streamIndex int) error {
 	ist, err := this.GetStream(streamIndex)
 	if err != nil {
 		return err
 	}
 
-	frameTs := Rescale(sec*1000, ist.TimeBase().AVR().Den, ist.TimeBase().AVR().Num) / 1000
+	frameTs := Rescale(sec*1000, int64(ist.TimeBase().AVR().Den), int64(ist.TimeBase().AVR().Num)) / 1000
 
 	if err := this.SeekFile(ist, frameTs, frameTs, C.AVSEEK_FLAG_FRAME); err != nil {
 		return err
